requirements.txt
# Core dependencies
requests>=2.28.1
python-dotenv>=0.21.0
click>=8.1.3
rich>=12.6.0
tqdm>=4.64.1

# OSINT dependencies
beautifulsoup4>=4.11.1
selenium>=4.5.0
webdriver-manager>=3.8.4
tweepy>=4.12.0
linkedin-api>=2.0.0
shodan>=1.28.0
censys>=2.1.8
whois>=0.9.27
geopy>=2.3.0
emailhunter>=1.0.0

# Steganography dependencies
Pillow>=9.3.0
stegano>=0.10.2
stepic>=0.5.0
pydub>=0.25.1
numpy>=1.23.4
opencv-python>=4.6.0.66

# Cryptography dependencies
cryptography>=38.0.1
pycryptodome>=3.15.0
bcrypt>=4.0.1
pyOpenSSL>=22.1.0
passlib>=1.7.4

# Tracking and reporting dependencies
pandas>=1.5.1
matplotlib>=3.6.2
seaborn>=0.12.1
networkx>=2.8.8
reportlab>=3.6.12
fpdf2>=2.5.7
graphviz>=0.20.1

# Generator and decoder dependencies
faker>=15.1.1
qrcode>=7.3.1
python-barcode>=0.14.0
base58>=2.1.1
base91>=1.0.1

# Web interface dependencies
flask>=2.2.2
flask-restful>=0.3.9
flask-cors>=3.0.10
flask-login>=0.6.2
gunicorn>=20.1.0

# Testing dependencies
pytest>=7.2.0
pytest-cov>=4.0.0
